+++ tasks/seq/korablev_v_quick_sort_simple_merge/func_tests/main.cpp
@@ -9,9 +9,9 @@
 TEST(korablev_v_quick_sort_simple_merge, test_sort_array) {
   const size_t array_size = 6;
   std::vector<size_t> in_size(1, array_size);
-  std::vector<double> input_data = {5.0, 3.0, 8.0, 6.0, 2.0, 7.0};
-  std::vector<double> expected_output = {2.0, 3.0, 5.0, 6.0, 7.0, 8.0};
-  std::vector<double> out(array_size, 0.0);
+  std::vector<int> input_data = {5, 3, 8, 6, 2, 7};
+  std::vector<int> expected_output = {2, 3, 5, 6, 7, 8};
+  std::vector<int> out(array_size, 0);
 
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(in_size.data()));
@@ -37,9 +37,9 @@ TEST(korablev_v_quick_sort_simple_merge, test_sort_array) {
 TEST(korablev_v_quick_sort_simple_merge, test_single_element) {
   const size_t array_size = 1;
   std::vector<size_t> in_size(1, array_size);
-  std::vector<double> input_data = {42.0};
-  std::vector<double> expected_output = {42.0};
-  std::vector<double> out(array_size, 0.0);
+  std::vector<int> input_data = {42};
+  std::vector<int> expected_output = {42};
+  std::vector<int> out(array_size, 0);
 
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(in_size.data()));
@@ -65,9 +65,9 @@ TEST(korablev_v_quick_sort_simple_merge, test_single_element) {
 TEST(korablev_v_quick_sort_simple_merge, test_all_equal_elements) {
   const size_t array_size = 4;
   std::vector<size_t> in_size(1, array_size);
-  std::vector<double> input_data = {7.0, 7.0, 7.0, 7.0};
-  std::vector<double> expected_output = {7.0, 7.0, 7.0, 7.0};
-  std::vector<double> out(array_size, 0.0);
+  std::vector<int> input_data = {7, 7, 7, 7};
+  std::vector<int> expected_output = {7, 7, 7, 7};
+  std::vector<int> out(array_size, 0);
 
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(in_size.data()));
@@ -93,9 +93,9 @@ TEST(korablev_v_quick_sort_simple_merge, test_all_equal_elements) {
 TEST(korablev_v_quick_sort_simple_merge, test_negative_and_positive) {
   const size_t array_size = 5;
   std::vector<size_t> in_size(1, array_size);
-  std::vector<double> input_data = {-3.0, 2.0, -1.0, 0.0, 1.0};
-  std::vector<double> expected_output = {-3.0, -1.0, 0.0, 1.0, 2.0};
-  std::vector<double> out(array_size, 0.0);
+  std::vector<int> input_data = {-3, 2, -1, 0, 1};
+  std::vector<int> expected_output = {-3, -1, 0, 1, 2};
+  std::vector<int> out(array_size, 0);
 
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(in_size.data()));
@@ -125,7 +125,7 @@ TEST(korablev_v_quick_sort_simple_merge, invalid_input_count) {
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(in_size.data()));
   taskDataSeq->inputs_count.emplace_back(in_size.size());
 
-  std::vector<double> out(6, 0.0);
+  std::vector<int> out(6, 0);
   taskDataSeq->outputs.emplace_back(reinterpret_cast<uint8_t*>(out.data()));
   taskDataSeq->outputs_count.emplace_back(out.size());
 
@@ -137,7 +137,7 @@ TEST(korablev_v_quick_sort_simple_merge, invalid_output_count) {
   std::shared_ptr<ppc::core::TaskData> taskDataSeq = std::make_shared<ppc::core::TaskData>();
 
   std::vector<size_t> in_size(1, 6);
-  std::vector<double> input_data = {5.0, 3.0, 8.0, 6.0, 2.0, 7.0};
+  std::vector<int> input_data = {5, 3, 8, 6, 2, 7};
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(in_size.data()));
   taskDataSeq->inputs_count.emplace_back(in_size.size());
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(input_data.data()));
@@ -154,11 +154,11 @@ TEST(korablev_v_quick_sort_simple_merge, invalid_negative_size) {
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(&invalid_size));
   taskDataSeq->inputs_count.emplace_back(1);
 
-  std::vector<double> input_data = {5.0, 3.0, 8.0, 6.0, 2.0, 7.0};
+  std::vector<int> input_data = {5, 3, 8, 6, 2, 7};
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(input_data.data()));
   taskDataSeq->inputs_count.emplace_back(input_data.size());
 
-  std::vector<double> out(6, 0.0);
+  std::vector<int> out(6, 0.0);
   taskDataSeq->outputs.emplace_back(reinterpret_cast<uint8_t*>(out.data()));
   taskDataSeq->outputs_count.emplace_back(out.size());
 
@@ -173,11 +173,11 @@ TEST(korablev_v_quick_sort_simple_merge, invalid_size_data_mismatch) {
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(&array_size));
   taskDataSeq->inputs_count.emplace_back(1);
 
-  std::vector<double> input_data = {5.0, 3.0, 8.0};
+  std::vector<int> input_data = {5, 3, 8};
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(input_data.data()));
   taskDataSeq->inputs_count.emplace_back(input_data.size());
 
-  std::vector<double> out(6, 0.0);
+  std::vector<int> out(6, 0);
   taskDataSeq->outputs.emplace_back(reinterpret_cast<uint8_t*>(out.data()));
   taskDataSeq->outputs_count.emplace_back(out.size());
 
@@ -192,11 +192,11 @@ TEST(korablev_v_quick_sort_simple_merge, invalid_output_size) {
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(&array_size));
   taskDataSeq->inputs_count.emplace_back(1);
 
-  std::vector<double> input_data = {5.0, 3.0, 8.0, 6.0, 2.0, 7.0};
+  std::vector<int> input_data = {5, 3, 8, 6, 2, 7};
   taskDataSeq->inputs.emplace_back(reinterpret_cast<uint8_t*>(input_data.data()));
   taskDataSeq->inputs_count.emplace_back(input_data.size());
 
-  std::vector<double> out(3, 0.0);
+  std::vector<int> out(3, 0);
   taskDataSeq->outputs.emplace_back(reinterpret_cast<uint8_t*>(out.data()));
   taskDataSeq->outputs_count.emplace_back(out.size());
